Move button component to src/components
