+++ docs/farming-&-staking/farming/advanced-cli/cli-tips.mdx
@@ -15,7 +15,7 @@ import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import Link from '@docusaurus/Link';
 import MDXComponents from '@theme-original/MDXComponents';
-import Button from '@site/src/theme/DocButton';
+import Button from '@site/src/components/Button';
 
 Whether you are an experienced farmer or new to the Autonomys Network, this section offers valuable insights to improve your experience and efficiency. We explore practical advice and hidden techniques that will help you optimize your farming setup and tackle common challenges effortlessly.
