+++ src/theme/MDXComponents.js
@@ -1,7 +1,7 @@
 import React from 'react';
 // Importing the original mapper + our components according to the Docusaurus doc
 import MDXComponents from '@theme-original/MDXComponents';
-import Button from '@site/src/theme/DocButton';
+import Button from '@site/src/components/Button';
 export default {
   // Reusing the default mapping
   ...MDXComponents,
